move to otel_thread_ctx module
