+++ src/native/lib.rs
@@ -14,6 +14,8 @@ mod ffe;
 mod http_client;
 mod library_config;
 mod log;
+#[cfg(target_os = "linux")]
+mod otel_thread_ctx;
 mod py_string;
 mod rand;
 mod shared_runtime;
@@ -55,10 +57,15 @@ fn _native(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<library_config::PyTracerMetadata>()?;
     m.add_class::<library_config::PyAnonymousFileHandle>()?;
     m.add_wrapped(wrap_pyfunction!(library_config::store_metadata))?;
+
     #[cfg(target_os = "linux")]
     {
-        m.add_wrapped(wrap_pyfunction!(library_config::update_otel_thread_context))?;
-        m.add_wrapped(wrap_pyfunction!(library_config::detach_otel_thread_context))?;
+        m.add_wrapped(wrap_pyfunction!(
+            otel_thread_ctx::update_otel_thread_context
+        ))?;
+        m.add_wrapped(wrap_pyfunction!(
+            otel_thread_ctx::detach_otel_thread_context
+        ))?;
     }
     shared_runtime::register_shared_runtime(m)?;
     data_pipeline::register_data_pipeline(m)?;
