+++ src/native/library_config.rs
@@ -7,9 +7,6 @@ use pyo3::prelude::*;
 use pyo3::types::PyDict;
 use pyo3::types::PyList;
 
-#[cfg(target_os = "linux")]
-use crate::span::SpanData;
-
 #[pyclass(name = "PyConfigurator", module = "ddtrace.internal._native")]
 pub struct PyConfigurator {
     configurator: Box<Configurator>,
@@ -142,34 +139,4 @@ pub fn store_metadata(data: &PyTracerMetadata) -> PyResult<PyAnonymousFileHandle
             Err(PyException::new_err(err_msg))
         }
     }
-}
-
-#[cfg(target_os = "linux")]
-#[pyfunction]
-pub fn update_otel_thread_context(
-    span: PyRef<'_, SpanData>,
-    local_root: Option<PyRef<'_, SpanData>>,
-) {
-    use libdd_otel_thread_ctx::linux::ThreadContext;
-
-    let local_root_span_id = if let Some(local_root) = local_root {
-        local_root.span_id
-    } else {
-        span.span_id
-    };
-
-    ThreadContext::update(
-        span.trace_id.to_be_bytes(),
-        span.span_id.to_be_bytes(),
-        local_root_span_id.to_be_bytes(),
-        &[],
-    );
-}
-
-#[cfg(target_os = "linux")]
-#[pyfunction]
-pub fn detach_otel_thread_context() {
-    use libdd_otel_thread_ctx::linux::ThreadContext;
-
-    ThreadContext::detach();
 }
